Fix contract for `read-markdown`.
Why didn't the unit tests catch this? I'm accustomed to using
`define/contract`, with which a `(module+ test)` will use and test the
contract. But now that I'm trying to use `(provide (contract-out))` --
because it's faster, and because it's the One True Way -- I'm
discovering that doesn't happen. Got bit that way, here.