Fix #18.
As of commit 3a9ce181 and the addition of `replace/match`, the href
portion of the link was being run through `intra-block` for further
parsing. Which is obviously wrong: It should be taken literally.

Give `replace/match` an opional argument, `intra-block?`, which is
either `#t` or a `(listof boolean?)`.  When `#t`, all matched elements
are run through `intra-block`, which was the old behavior. When a
`(listof boolean?)`, it says which of the matched elements should be
run through intra-block: For example, `'(#t #f) ` means to run the
first one through intra-block but not the second.

Also fix parallel bug with ![]() image links.