Ensure `html->str` supplies close tags for non-void elements.
Move `void-element?` from parse.rkt to its own file so we can use it
here, too.

Even if a non-void element lacks contents, it MUST supply a close
tag. Otherwise the markdown parser will not treat it as valid HTML.

Would a better way to do this be to modify the redex grammar for
non-void elements in such a way that randomly-generated examples
_always_ have content? Perhaps, but (a) I don't yet understand how to
do that, and (b) I'm not sure that's really correct grammar. A
non-void element is permitted to have no contents. The point is that,
regardless of its contents, it must _always_ have the closing
</tag>. I really think this is a matter for html->str, at the
"boundary" of changing sexprs to HTML text.