Fix approach to footnotes. Closes #19.
Untangle reference links and footnotes; different requirements.

Footnote uses and footnote definitions may appear in any order in the
source. (Including all of the definitions before any of the uses,
alternating uses and definitions, or whatever order.)

Footnotes are numbered by the order the footnote uses appear in the
source, of course.

Definitions should be gathered into a single `div` at the end, using
an `ol` to number them automatically.

Definitions should appear in order based on the footnote number (not
on the order of the definitions in the source).