Change random-test.rkt to generate "tokens".
Previous version of this I used completely random characters. However
we're much more likely to generate something that could break the
parser if we randomly choose among "tokens" that mean something
special in markdown (as well as from plain words). This results in
something that looks like it could be markdown -- both to a human and
a parser.

Soon after using this I already discovered a bug in the fenced code
block grammar that would cause a syntax error. And I also discovered
an input that causes the parser never to terminate. So that's good
news, at least from one perspetive.