Revise approach to fixing issue #27.
Require users of `parse-markdown` to pass \n terminated lines, and
suggest they use `#:mode 'text` when opening markdown files (which
does automatic \r\n -> \n conversion on Windows). And, update
`test.rkt` to follow this advice! :)

In fact, revert commit 5f26ea36 which replaced uses of $newline with
$eol. That turned out to be insufficient. There are also numerous
examples of things like `(oneOf "\n")` -- a $newline "by other means"
-- which would need to be changed to `(oneOf "\r\n")`.  Although easy
in some cases, not so simple in others. Plus, it would litter the code
with this.

The parser is clearest when it deals with the concept of "line
termination". It's not really the level at which you want to deal with
varying conventions for representing line termination. Stipulate it's
\n for the parser, and let `#:mode 'text` do what it's intended to do.

Why revert the commit? I didn't like the idea of $eol being used in
many places, while the moral equivalent of $newline remained in many
others. Although that would have "worked", today, who knows what
confusion it might have caused in the future.