Add dep on `sexp-diff` pkg. Bump to v0.8.
Use sexp-diff for vastly better output from suite.rkt. Some of the
xexprs are very large. This pinpoints the differences.

Speaking of suite.rkt, move its contents inside a `test-suite` module,
so not part of main load image. And not in a `test` submodule because
two of the tests still fail. (They fail justifiably IMHO, but they
fail. Among other things, this would bugger the Travis CI build
status.)