Block images (unless strict mode).
An image by itself on a line outputs the <img> nested with a
<p class="caption"> in a <div class="figure">.

Not only does this allow the `alt` attribute to serve as a visible
caption, it allows CSS to style the <div> (as well as the caption
<p>). This should make images in e.g. blog posts more attractive,
because you can use CSS to control the horizontal align and margins,
and do so consistently.

While exploring reference links using Babelmark2, I noticed that
Pandoc non-strict does this, and I like it.

Images with other text on the same line (and all images in strict
mode) output just an <img> element as usual.