Changes to emph and strong.
random-test.rkt was very slow since commit 42cbb412. I think it was
due to unnecessary runtime function creation.

But also, the correct parse that fixes #30 is not going to deal so
well with deliberately fake emph or strong. So I concluded that
random-test.rkt was being too "provocative", and dialed that back.