Don't expand tabs in pre elements.
On reflection I think the correct thing to do is leave them as-is.

If someone wants to filter them, that's properly a job for something
like Pygments post-processing.

Plus you probably don't want to filter them. Consider the example of a
code block containing a makefile, where tabs are significant (IIUC).
Someone should be able to copy and paste that example.